chore: off-by-one error on unvoided tx scenario
