+++ packages/daemon/__tests__/integration/balances.test.ts
@@ -10,15 +10,13 @@ import { SyncMachine } from '../../src/machines';
 import { interpret } from 'xstate';
 import { getDbConnection } from '../../src/db';
 import { Connection } from 'mysql2/promise';
-import { 
-  cleanDatabase, 
-  fetchAddressBalances, 
-  transitionUntilEvent, 
-  validateBalances, 
-  validateBalanceDistribution,
+import {
+  cleanDatabase,
+  fetchAddressBalances,
+  transitionUntilEvent,
+  validateBalances,
   performVoidingConsistencyChecks,
   validateVoidingConsistency,
-  printVoidingConsistencyReport
 } from './utils';
 import unvoidedScenarioBalances from './scenario_configs/unvoided_transactions.balances';
 import reorgScenarioBalances from './scenario_configs/reorg.balances';
@@ -81,28 +79,24 @@ getConfig.mockReturnValue({
   DB_PORT,
 });
 
-// Use a single mysql connection for all tests
 let mysql: Connection;
-beforeAll(async () => {
-  try {
-    mysql = await getDbConnection();
-  } catch (e) {
-    console.error('Failed to establish db connection', e);
-    throw e;
-  }
-});
 
-beforeEach(async () => {
+beforeAll(async () => {
+  mysql = await getDbConnection();
   await cleanDatabase(mysql);
 });
 
-describe('unvoided transaction scenario', () => {
-  beforeAll(() => {
-    jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
-  });
+afterAll(async () => {
+  jest.resetAllMocks();
+  if (mysql && 'release' in mysql) {
+    // @ts-expect-error - pooled connection has release method
+    await mysql.release();
+  }
+});
 
-  afterAll(() => {
-    jest.resetAllMocks();
+describe('unvoided transaction scenario', () => {
+  beforeAll(async () => {
+    jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -130,13 +124,14 @@ describe('unvoided transaction scenario', () => {
     // @ts-expect-error
     await transitionUntilEvent(mysql, machine, UNVOIDED_SCENARIO_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
-    expect(validateBalances(addressBalances, unvoidedScenarioBalances));
+    await validateBalances(addressBalances, unvoidedScenarioBalances);
   });
 });
 
 describe('reorg scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -164,13 +159,14 @@ describe('reorg scenario', () => {
     // @ts-expect-error
     await transitionUntilEvent(mysql, machine, REORG_SCENARIO_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
-    expect(validateBalances(addressBalances, reorgScenarioBalances));
+    await validateBalances(addressBalances, reorgScenarioBalances);
   });
 });
 
 describe('single chain blocks and transactions scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -198,13 +194,14 @@ describe('single chain blocks and transactions scenario', () => {
     // @ts-expect-error
     await transitionUntilEvent(mysql, machine, SINGLE_CHAIN_BLOCKS_AND_TRANSACTIONS_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
-    expect(validateBalances(addressBalances, singleChainBlocksAndTransactionsBalances));
+    await validateBalances(addressBalances, singleChainBlocksAndTransactionsBalances);
   });
 });
 
 describe('invalid mempool transactions scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -232,13 +229,14 @@ describe('invalid mempool transactions scenario', () => {
     // @ts-expect-error
     await transitionUntilEvent(mysql, machine, INVALID_MEMPOOL_TRANSACTION_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
-    expect(validateBalances(addressBalances, invalidMempoolBalances));
+    await validateBalances(addressBalances, invalidMempoolBalances);
   });
 });
 
 describe('custom script scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -266,13 +264,14 @@ describe('custom script scenario', () => {
     // @ts-expect-error
     await transitionUntilEvent(mysql, machine, CUSTOM_SCRIPT_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
-    expect(validateBalances(addressBalances, customScriptBalances));
+    await validateBalances(addressBalances, customScriptBalances);
   });
 });
 
 describe('empty script scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -301,13 +300,14 @@ describe('empty script scenario', () => {
     await transitionUntilEvent(mysql, machine, EMPTY_SCRIPT_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
 
-    expect(validateBalances(addressBalances, emptyScriptBalances));
+    await validateBalances(addressBalances, emptyScriptBalances);
   });
 });
 
 describe('nc events scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match', async () => {
@@ -336,14 +336,14 @@ describe('nc events scenario', () => {
     await transitionUntilEvent(mysql, machine, NC_EVENTS_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
 
-    // @ts-ignore
-    expect(validateBalances(addressBalances, ncEventsBalances));
+    await validateBalances(addressBalances, ncEventsBalances);
   });
 });
 
 describe('transaction voiding chain scenario', () => {
-  beforeAll(() => {
+  beforeAll(async () => {
     jest.spyOn(Services, 'fetchMinRewardBlocks').mockImplementation(async () => 300);
+    await cleanDatabase(mysql);
   });
 
   it('should do a full sync and the balances should match after voiding chain', async () => {
@@ -372,7 +372,28 @@ describe('transaction voiding chain scenario', () => {
     await transitionUntilEvent(mysql, machine, TRANSACTION_VOIDING_CHAIN_LAST_EVENT);
     const addressBalances = await fetchAddressBalances(mysql);
 
-    // Validate balance distribution
-    validateBalanceDistribution(addressBalances, transactionVoidingChainBalances);
+    await validateBalances(addressBalances, transactionVoidingChainBalances);
+
+    // Validate transaction voiding consistency
+    const voidingChecks = await performVoidingConsistencyChecks(mysql, {
+      transactions: [
+        { txId: '404eeba6f3658028722e665684c42a0c28c3a44b190426971e012c5030bf1903', expectedVoided: false }, // tx1
+        { txId: '4412a1f718b51c054193cc0df2d1dd9d16e82684be17760d7169aa6fb22d5ea2', expectedVoided: false }, // tx2
+        { txId: '06b20fd4258ae965137203d4e1fd7df7b69e775e5d3f4a4568d1161343b91f02', expectedVoided: true },  // spending_tx (voided)
+        { txId: 'ada5e728ffd680238306b510899629099d6bbb58a8811b042c249a236f9640cc', expectedVoided: false }, // voiding_tx
+      ],
+      utxos: [
+        // Spending transaction (voided) UTXOs should be marked as voided
+        { txId: '06b20fd4258ae965137203d4e1fd7df7b69e775e5d3f4a4568d1161343b91f02', index: 0, expectedValue: 5900, expectedVoided: true, expectedSpentBy: null },
+        { txId: '06b20fd4258ae965137203d4e1fd7df7b69e775e5d3f4a4568d1161343b91f02', index: 1, expectedValue: 500, expectedVoided: true, expectedSpentBy: null },
+
+        // Voiding transaction (valid) UTXOs should not be voided
+        { txId: 'ada5e728ffd680238306b510899629099d6bbb58a8811b042c249a236f9640cc', index: 0, expectedValue: 5900, expectedVoided: false, expectedSpentBy: null },
+        { txId: 'ada5e728ffd680238306b510899629099d6bbb58a8811b042c249a236f9640cc', index: 1, expectedValue: 500, expectedVoided: false, expectedSpentBy: null },
+      ],
+    });
+
+    // Validate consistency
+    validateVoidingConsistency(voidingChecks);
   }, 30000); // 30 second timeout for transaction voiding chain test
 });
